feat: BaseTimeEntity 생성
