--- src/main/java/reviewers/customer/assignmentGrade/controller/AssignmentController.java
@@ -1,4 +1,0 @@
-package reviewers.customer.assignmentGrade.controller;
-
-public class AssignmentController {
-}
