--- src/main/java/reviewers/customer/assignmentGrade/entity/AssignmentGrade.java
@@ -1,4 +1,0 @@
-package reviewers.customer.assignmentGrade.entity;
-
-public class AssignmentGrade {
-}
