--- src/main/java/reviewers/customer/assignmentGrade/service/AssignmentService.java
@@ -1,4 +1,0 @@
-package reviewers.customer.assignmentGrade.service;
-
-public class AssignmentService {
-}
