+++ src/main/java/reviewers/customer/domain/assignmentGrade/controller/AssignmentController.java
@@ -1,0 +1,4 @@
+package reviewers.customer.domain.assignmentGrade.controller;
+
+public class AssignmentController {
+}
