+++ src/main/java/reviewers/customer/domain/assignmentGrade/entity/AssignmentGrade.java
@@ -1,0 +1,4 @@
+package reviewers.customer.domain.assignmentGrade.entity;
+
+public class AssignmentGrade {
+}
