+++ src/main/java/reviewers/customer/domain/assignmentGrade/service/AssignmentService.java
@@ -1,0 +1,4 @@
+package reviewers.customer.domain.assignmentGrade.service;
+
+public class AssignmentService {
+}
