+++ src/main/java/reviewers/customer/domain/quizGrade/controller/QuizController.java
@@ -1,0 +1,4 @@
+package reviewers.customer.domain.quizGrade.controller;
+
+public class QuizController {
+}
