+++ src/main/java/reviewers/customer/domain/quizGrade/entity/QuizGrade.java
@@ -1,0 +1,4 @@
+package reviewers.customer.domain.quizGrade.entity;
+
+public class QuizGrade {
+}
