+++ src/main/java/reviewers/customer/domain/quizGrade/service/QuizService.java
@@ -1,0 +1,4 @@
+package reviewers.customer.domain.quizGrade.service;
+
+public class QuizService {
+}
