--- src/main/java/reviewers/customer/quizGrade/controller/QuizController.java
@@ -1,4 +1,0 @@
-package reviewers.customer.quizGrade.controller;
-
-public class QuizController {
-}
