--- src/main/java/reviewers/customer/quizGrade/entity/QuizGrade.java
@@ -1,4 +1,0 @@
-package reviewers.customer.quizGrade.entity;
-
-public class QuizGrade {
-}
