--- src/main/java/reviewers/customer/quizGrade/service/QuizService.java
@@ -1,4 +1,0 @@
-package reviewers.customer.quizGrade.service;
-
-public class QuizService {
-}
